backend/requirements.txt
Django==2.1.7
django-webpack-loader==0.6.0
djangorestframework==3.9.1
pytz==2018.9
Pillow==5.4.1
